feat(inventory-uom-category): Refactor - Add new uom category template strcuture
