+++ moduerp/inventory/static/js/uom/utils/get_cookie.js
@@ -1,4 +1,4 @@
-function getCookie(name) {
+export function getCookie(name) {
     let cookieValue = null;
     if (document.cookie && document.cookie !== '') {
         const cookies = document.cookie.split(';');
