+++ moduerp/inventory/templates/uom/components/uom_category_new_form.html
@@ -1,0 +1,86 @@
+{% load static %}
+
+<div class="flex-grow-1 overflow-auto m-0 w-100 h-100 p-3">
+  <div class="p-3 bg-dark rounded">
+    <!-- UoM Category name -->
+    <div class="row mb-2">
+      <div class="col-12 col-md-2 fw-semibold text-white my-2">
+        Unit of Measure Category: 
+      </div>
+      <div class="col-12 col-md-5 d-flex flex-row gap-2 my-2">
+        <input type="text" id="uom-category-name-input" 
+          class="form-control form-control-sm border-0 border-bottom border-blue-grey-5 rounded-0 bg-dark text-white fw-light placeholder-blue-grey-5" 
+          placeholder="Enter UoM Category">
+      </div>
+    </div>
+
+    <!-- UoM Table -->
+    <table id="uom-table" class="table-hover align-middle h-100 w-100">
+      <thead class="border-top border-bottom border-white bg-blue-grey-9" style="position: sticky; top: 0; z-index: 10;">
+        <tr>
+          <th class="text-start p-2" style="width: 35%;">Unit of Measure</th>
+          <th class="text-start p-2" style="width: 20%;">Type</th>
+          <th class="text-end p-2" style="width: 20%;">Ratio</th>
+          <th class="text-center p-2" style="width: 20%;">Default</th>
+          <th class="text-center p-2" style="width: 5%;"></th>
+        </tr>
+      </thead>
+      <tbody>
+        <!-- Reference UoM -->
+        <tr class="border-bottom border-bottom-1 border-blue-grey-5">
+          <td class="text-start p-1">
+            <input type="text" name="uom-name-input"
+                class="form-control form-control-sm bg-dark text-white fw-light placeholder-blue-grey-5 border-0" 
+                placeholder="Unit of Measure">
+          </td>
+          <td class="text-start p-1">
+            <select name="uom-type-select" class="form-select form-select-sm bg-dark text-white border-0" disabled>
+              <option value="reference" selected>Reference Unit</option>
+              <option value="bigger">Bigger than Reference</option>
+              <option value="smaller">Smaller than Reference</option>
+            </select>
+          </td>
+          <td class="text-end p-1">
+            <input type="number" name="uom-ratio-input"
+                class="form-control form-control-sm bg-dark text-white placeholder-blue-grey-5 border-0"
+                placeholder="Ratio" style="text-align: right;" value="1.0" readonly>
+          </td>
+          <td class="text-center p-1"><input name="uom-default-input" type="checkbox" class="form-check-input bg-dark" checked></td>
+          <td class="text-center p-1"></td>
+        </tr>
+      </tbody>
+    </table>
+
+    <!-- Add UoM button -->
+    <button id="add-uom-btn" type="button" class="btn btn-indigo-5 btn-sm mt-3">New Unit of Measure</button>
+  </div>
+</div>
+
+<template id="uom-row-template">
+      <tr class="border-bottom border-bottom-1 border-blue-grey-5">
+        <td class="text-start p-1">
+            <input type="text" name="uom-name-input" 
+                class="form-control form-control-sm bg-dark text-white fw-light placeholder-blue-grey-5 border-0" 
+                placeholder="Unit of Measure">
+        </td>
+        <td class="text-start p-1">
+          <select name="uom-type-select" class="form-select form-select-sm bg-dark text-white fw-light border-0">
+            <option value="bigger">Bigger than Reference</option>
+            <option value="smaller">Smaller than Reference</option>
+          </select>
+        </td>
+        <td class="text-end p-1">
+            <input type="number" name="uom-ratio-input"
+                class="form-control form-control-sm bg-dark text-white fw-light placeholder-blue-grey-5 border-0"
+                placeholder="Ratio" style="text-align: right;">
+        </td>
+        <td class="text-center p-1"><input type="checkbox" name="uom-default-input" class="form-check-input bg-dark"></td>
+        <td class="text-center p-1">
+            <button type="button" class="remove-btn btn btn-pink-5 btn-sm delete-filter">
+                <img src="{% static 'icons/trash.svg' %}" alt="trash" height="20">
+            </button>
+        </td>
+      </tr>
+</template>
+
+<script type="module" src="{% static 'js/uom/components/uom_category_new_form.js' %}"></script>
