+++ moduerp/templates/components/toast.html
@@ -12,4 +12,3 @@
     </div>
   </div>
 </template>
-<script src="{% static 'js/components/toast.js' %}"></script>
